requirements.txt
# LLM API clients
openai>=1.30.0
mistralai>=1.0.0
together>=1.2.0

# Embeddings
sentence-transformers>=2.6.0

# Dimensionality reduction & clustering
umap-learn>=0.5.6
scikit-learn>=1.4.0

# Topic modeling
bertopic>=0.16.0

# Statistical analysis
statsmodels>=0.14.0
pingouin>=0.5.4
scipy>=1.11.0

# NLP utilities
spacy>=3.7.0
nltk>=3.8.0

# Transformers (GPT-2 XL co-occurrence baseline)
transformers>=4.40.0
torch>=2.2.0

# Data handling
pandas>=2.0.0
numpy>=1.26.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Environment
python-dotenv>=1.0.0
tqdm>=4.66.0
